+++ changelog.md
@@ -1,3 +1,5 @@
 # 12-09-2022 / Beta v.0.5 :
   * Add Changelog!
   * update readme
+  * Add Export to CSV
+  * Add Export to Excel (xlsx format)